vdl: move the vdlgen.JSON to javascript/json
This was suggested in #29 by @tatatodd.
